requirements.txt
boto3~=1.35.27
PyYAML~=6.0.2
botocore~=1.35.27